+++ app/javascript/components/HeaderNav.js
@@ -9,7 +9,7 @@ import NavDropdown from 'react-bootstrap/NavDropdown';
 import { useHistory } from 'react-router-dom';
 
 import {
-  updateUser, allCommunities, allDoctors,
+  updateUser, allCommunities, allDoctors, signOutUser,
 } from '../redux/actions';
 
 const HeaderNav = ({ dispatch, user }) => {
@@ -41,6 +41,7 @@ const HeaderNav = ({ dispatch, user }) => {
         dispatch(updateUser(data));
       }).catch(err => console.log(err));
   }, [history]);
+
   useEffect(() => {
     // const url = 'http://localhost:3000/user';
     // const url = 'https://obscure-island-28750.herokuapp.com/user';
@@ -88,7 +89,7 @@ const HeaderNav = ({ dispatch, user }) => {
     ? (
       <div style={{ display: 'flex', justifyContent: 'space-between' }}>
         <Nav.Link onClick={() => history.push(`/users/${user.name}`)}>{user.name}</Nav.Link>
-        <Nav.Link className="signOutBtn" onClick={() => { dispatch(updateUser({ name: '' })); localStorage.token = ''; }}>Sign out</Nav.Link>
+        <Nav.Link className="signOutBtn" onClick={() => { dispatch(signOutUser({ name: '' })); localStorage.token = ''; }}>Sign out</Nav.Link>
       </div>
     )
     : (
@@ -127,7 +128,7 @@ const HeaderNav = ({ dispatch, user }) => {
           </Nav>
         </Navbar.Collapse>
         {user.name !== ''
-          ? <Nav.Link className="bigScreenSignOutBtn" onClick={() => { dispatch(updateUser({ name: '' })); localStorage.token = ''; }}>Sign out</Nav.Link>
+          ? <Nav.Link className="bigScreenSignOutBtn" onClick={() => { dispatch(signOutUser({ name: '' })); localStorage.token = ''; }}>Sign out</Nav.Link>
           : null}
       </Navbar>
     </div>
