+++ app/javascript/redux/store.js
@@ -1,7 +1,7 @@
 import { combineReducers, createStore } from 'redux';
 
 import {
-  SUBMIT, COMMUNITY, COMMUNITIES, TAB, DOCTORS, DOCTOR, MYDOCTORS,
+  SUBMIT, COMMUNITY, COMMUNITIES, TAB, DOCTORS, DOCTOR, MYDOCTORS, SIGNOUT,
 } from './actions';
 
 const initialUser = { name: '' };
@@ -12,9 +12,38 @@ const userReducer = (state = initialUser, action) => {
   switch (action.type) {
     case SUBMIT:
       return {
-        user: action.input,
+        user: {
+          id: action.input.id,
+          name: action.input.name,
+          email: action.input.email,
+          age: action.input.age,
+          dob: action.input.dob,
+          bloodSugar: JSON.parse(action.input.blood_sugar),
+          diastolic: JSON.parse(action.input.diastolic),
+          height: JSON.parse(action.input.height),
+          pulse: JSON.parse(action.input.pulse),
+          systolic: JSON.parse(action.input.systolic),
+          temperature: JSON.parse(action.input.temperature),
+          weight: JSON.parse(action.input.weight),
+        },
       };
-
+    case SIGNOUT:
+      return {
+        user: {
+          name: action.input.name,
+          id: 0,
+          email: '',
+          age: '',
+          dob: '',
+          bloodSugar: {},
+          diastolic: {},
+          height: {},
+          pulse: {},
+          systolic: {},
+          temperature: {},
+          weight: { measurements: '' },
+        },
+      };
     default:
       return state;
   }
